+++ code_producers/src/c_elements/c_code_generator.rs
@@ -964,7 +964,7 @@ pub fn generate_fr_hpp_file(c_folder: &PathBuf, prime: &String, producer: &CProd
             "pallas" => include_str!("pallas/fr.hpp"),
             "vesta" => include_str!("vesta/fr.hpp"),
             "secq256r1" => include_str!("secq256r1/fr.hpp"),
-            "bls12-377" => include_str!("bls12-377/fr.hpp"),
+            "bls12377" => include_str!("bls12377/fr.hpp"),
             _ => unreachable!(),
         };
         for line in file.lines() {
@@ -1078,7 +1078,7 @@ pub fn generate_fr_cpp_file(c_folder: &PathBuf, prime: &String,  producer: &CPro
                 "pallas" => include_str!("pallas/fr.cpp"),
                 "vesta" => include_str!("vesta/fr.cpp"),
                 "secq256r1" => include_str!("secq256r1/fr.cpp"),
-                "bls12-377" => include_str!("bls12-377/fr.cpp"),
+                "bls12377" => include_str!("bls12377/fr.cpp"),
                 _ => unreachable!(),
             };
             for line in file.lines() {
@@ -1126,7 +1126,7 @@ pub fn generate_fr_asm_file(c_folder: &PathBuf, prime: &String, producer: &CProd
             "pallas" => include_str!("pallas/fr.asm"),
             "vesta" => include_str!("vesta/fr.asm"),
             "secq256r1" => include_str!("secq256r1/fr.asm"),
-            "bls12-377" => include_str!("bls12-377/fr.asm"),
+            "bls12377" => include_str!("bls12377/fr.asm"),
             
             _ => unreachable!(),
         };