+++ code_producers/src/wasm_elements/wasm_code_generator.rs
@@ -1718,7 +1718,7 @@ pub fn fr_types(prime: &String) -> Vec<WasmInstruction> {
         "pallas" => include_str!("pallas/fr-types.wat"),
         "vesta" => include_str!("vesta/fr-types.wat"),
         "secq256r1" => include_str!("secq256r1/fr-types.wat"),
-        "bls12-377" => include_str!("bls12-377/fr-types.wat"),
+        "bls12377" => include_str!("bls12377/fr-types.wat"),
         _ => unreachable!(),
     };    
     for line in file.lines() {
@@ -1737,7 +1737,7 @@ pub fn fr_data(prime: &String) -> Vec<WasmInstruction> {
         "pallas" => include_str!("pallas/fr-data.wat"),
         "vesta" => include_str!("vesta/fr-data.wat"),
         "secq256r1" => include_str!("secq256r1/fr-data.wat"),
-        "bls12-377" => include_str!("bls12-377/fr-data.wat"),
+        "bls12377" => include_str!("bls12377/fr-data.wat"),
         _ => unreachable!(),
     };    
     for line in file.lines() {
@@ -1755,7 +1755,7 @@ pub fn fr_code(prime: &String) -> Vec<WasmInstruction> {
         "pallas" => include_str!("pallas/fr-code.wat"),
         "vesta" => include_str!("vesta/fr-code.wat"),
         "secq256r1" => include_str!("secq256r1/fr-code.wat"),
-        "bls12-377" => include_str!("bls12-377/fr-code.wat"),
+        "bls12377" => include_str!("bls12377/fr-code.wat"),
         _ => unreachable!(),
     };    
     for line in file.lines() {